style: rename
 - Rename Employees.ts to EmployeesStore.ts to indicate that it's store
 - same style with other store files
